-Added updated version of data generation related files -Fixed problems in SBO_ANCAS -Still some problems with running SBO ANCAS with high precision - TBD solve this problem

+++ Code/Implementations/TcaCalculation.h
@@ -10,8 +10,8 @@ namespace TcaCalculation
 	/// Time of Closest Approach (TCA)
 	/// </summary>
 	struct TCA {
-		long double time;
-		long double distance;
+		double time;
+		double distance;
 		int numberOfPoints;
 	};
